requirements.txt
requests==2.32.3
streamlit==1.45.0
uvicorn==0.23.2
python-dotenv==1.0.1
httpx==0.25.2
groq==0.11.0
